move IndustryData.ts to data directory
