+++ src/Corporation/ui/IndustryDescription.tsx
@@ -5,7 +5,7 @@ import Typography from "@mui/material/Typography";
 import { MoneyCost } from "./MoneyCost";
 import { Corporation } from "../Corporation";
 import { IndustryType } from "../data/Enums";
-import { IndustriesData } from "../IndustryData";
+import { IndustriesData } from "../data/IndustryData";
 
 interface IProps {
   industry: IndustryType;
