+++ src/Corporation/ui/NewDivisionTab.tsx
@@ -7,7 +7,7 @@ import TextField from "@mui/material/TextField";
 import Typography from "@mui/material/Typography";
 
 import { dialogBoxCreate } from "../../ui/React/DialogBox";
-import { IndustriesData } from "../IndustryData";
+import { IndustriesData } from "../data/IndustryData";
 import { IndustryType } from "../data/Enums";
 import { useCorporation } from "./Context";
 import { NewDivision } from "../Actions";
